auto import from //depot/cupcake/@132589

+++ tools/ddms/libs/ddmuilib/src/com/android/ddmuilib/NativeHeapPanel.java
@@ -237,7 +237,7 @@ public final class NativeHeapPanel extends BaseHeapPanel {
      */
     private HashMap<Long, NativeStackCallInfo> mSourceCache =
         new HashMap<Long,NativeStackCallInfo>();
-    private long mTotalSize;
+    private int mTotalSize;
     private Button mSaveButton;
     private Button mSymbolsButton;
